requirements.txt
selenium
streamlit
streamlit_ace
gemini-pro-api
google-generativeai
streamlit_authenticator
pdf2image
python-dotenv
plotly
pytesseract
webdriver-manager
crewai
crewai_tools
python-docx 
openai 
fpdf
pdf2image
plotly
pyautogui